+++ tests/test_geo_utils.py
@@ -4,8 +4,22 @@ Test geo utils.
 import numpy as np
 from osgeo import gdal
 from qgis.core import QgsCoordinateTransform
-from threedi_results_analysis.tests.utilities import ensure_qgis_app_is_initialized, create_test_raster
-from threedi_results_analysis.utils.geo_utils import get_coord_transformation_instance, get_authority_code
+from threedi_results_analysis.tests.utilities import (
+    ensure_qgis_app_is_initialized,
+    create_test_raster,
+    create_test_raster_with_defaults,
+    OFFSET_X,
+    OFFSET_Y,
+    OUT_DIR,
+)
+from threedi_results_analysis.utils.geo_utils import (
+    get_coord_transformation_instance,
+    get_authority_code,
+    get_extent,
+    replace_no_data_values,
+    get_shared_extent,
+    get_arrays,
+)
 
 import pytest
 
@@ -56,3 +70,53 @@ def test_get_authority_code(tmpdir, epsg_code, expected_epsg_code):
     raster_file = create_test_raster(str(outpath), np.zeros((10, 10)), origin_x=0, origin_y=0, epsg_code=epsg_code)
     raster = gdal.Open(raster_file)
     assert get_authority_code(raster) == expected_epsg_code
+
+
+def get_raster(array_data, origin_x, origin_y, no_data_value=-9999):
+    raster_file = create_test_raster_with_defaults("temp.tif", array_data, origin_x, origin_y, no_data_value=no_data_value)
+    return gdal.Open(raster_file)
+
+
+def test_get_extent(raster_zeros):
+    extent = get_extent(raster_zeros)
+    np.testing.assert_array_equal(extent, (OFFSET_X, OFFSET_Y, OFFSET_X + 100, OFFSET_Y - 100))
+
+
+def test_get_shared_extent(raster_zeros):
+    raster_share = get_raster(np.zeros((10, 10)), OFFSET_X + 50, OFFSET_Y + 50)
+    shared_extent = get_shared_extent(get_extent(raster_zeros), get_extent(raster_share))
+    np.testing.assert_array_equal(shared_extent, (OFFSET_X + 50, OFFSET_Y, OFFSET_X + 100, OFFSET_Y - 50))
+
+
+class TestGetArrays:
+
+    def test_extent_mismatch(self):
+        raster1 = get_raster(np.ones((10, 10)), origin_x=OFFSET_X, origin_y=OFFSET_Y)
+        data2 = np.arange(100).reshape(10, 10)
+        raster2 = get_raster(data2, origin_x=OFFSET_X + 50, origin_y=OFFSET_Y + 50)
+        array1, array2, gt = get_arrays(raster1, raster2)
+        np.testing.assert_array_equal(array1, np.ones((5, 5)))
+        np.testing.assert_array_equal(array2, data2[5:, 0:5])
+        np.testing.assert_array_equal(gt, (OFFSET_X + 50, 10, 0, OFFSET_Y, 0, -10))
+
+    def test_extent_match(self):
+        raster1 = get_raster(np.ones((10, 10)), origin_x=OFFSET_X, origin_y=OFFSET_Y)
+        data2 = np.arange(100).reshape(10, 10)
+        raster2 = get_raster(data2, origin_x=OFFSET_X, origin_y=OFFSET_Y)
+        array1, array2, gt = get_arrays(raster1, raster2)
+        np.testing.assert_array_equal(array1, np.ones((10, 10)))
+        np.testing.assert_array_equal(array2, data2)
+        np.testing.assert_array_equal(gt, (OFFSET_X, 10, 0, OFFSET_Y, 0, -10))
+
+
+def test_replace_no_data_values():
+    data = np.ones((10, 10))
+    data[1, 1] = -9999
+    data[3, 3] = -1337
+    raster = get_raster(np.ones((10, 10)), origin_x=OFFSET_X, origin_y=OFFSET_Y, no_data_value=-1337)
+    replace_no_data_values(raster, data)
+    expected_data = np.ones((10, 10))
+    expected_data[1, 1] = -9999
+    expected_data[3, 3] = 0
+    np.testing.assert_array_equal(data, expected_data)
+
